refactor: rename Generation → ResponsePipeline namespace + types
Mechanical rename to align code with the canonical Response pipeline
vocabulary established in CONTEXT.md. No behavior change.

- Move backend/Services/Generation/ → backend/Services/ResponsePipeline/
- Namespace PartyTown.Services.Generation → PartyTown.Services.ResponsePipeline
- GenerationSession → SpeakingSession; GenerationResult → SpeakingResult
- InFlightPhase.Generation → InFlightPhase.Speaking
- Update CONTEXT.md note to reflect post-rename surface

Leaves PersonaDecisionService / PersonaSalienceService / PersonaEmoteService
type names and Generation* log/tracing tags untouched.

Closes #59
