+++ backend-test/PersonaSalienceServiceTest.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
 using PartyTown.Grains.Generation;
 using PartyTown.Model;
-using PartyTown.Services.Generation;
+using PartyTown.Services.ResponsePipeline;
 
 namespace BackendTest;
