+++ backend/Services/ResponsePipeline/Prompts/DefaultCharacters/Denise.md
